passing value in main using by pointer in C program

+++ Call by Reference in C program/problem 1/main.c
@@ -4,7 +4,7 @@
 void SumAndAvg(int a, int b, int *sum, float *avg)
 {
     *sum = a + b;
-    *avg = *sum / 2;
+    *avg = (float)*sum / 2;
 }
 
 int main()
